Moved range representer into serialise
This is where all other custom yaml dumpers are
