+++ datacube/utils/serialise.py
@@ -13,6 +13,7 @@ import numpy
 import yaml
 
 from datacube.utils.documents import transform_object_tree
+from datacube.model._base import Range
 
 
 class SafeDatacubeDumper(yaml.SafeDumper):  # pylint: disable=too-many-ancestors
@@ -27,8 +28,25 @@ def _reduced_accuracy_decimal_representer(dumper: yaml.Dumper, data: Decimal) ->
     return dumper.represent_float(float(data))
 
 
+def _range_representer(dumper: yaml.Dumper, data: Range) -> yaml.Node:
+    begin, end = data
+
+    # pyyaml doesn't output timestamps in flow style as timestamps(?)
+    if isinstance(begin, datetime):
+        begin = begin.isoformat()
+    if isinstance(end, datetime):
+        end = end.isoformat()
+
+    return dumper.represent_mapping(
+        yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG,
+        (('begin', begin), ('end', end)),
+        flow_style=True
+    )
+
+
 SafeDatacubeDumper.add_representer(OrderedDict, _dict_representer)
 SafeDatacubeDumper.add_representer(Decimal, _reduced_accuracy_decimal_representer)
+SafeDatacubeDumper.add_representer(Range, _range_representer)
 
 
 def jsonify_document(doc):
